backend/requirements.txt
ultralytics
opencv-python
numpy
pillow
pyyaml
torch
torchvision
requests
albumentations
roboflow
flask
flask-cors
pandas
matplotlib
pymongo
python-dotenv
scikit-learn
seaborn
